change command palette to use moderation icon from ee
