+++ frontend/src/metabase/palette/hooks/useCommandPalette.tsx
@@ -26,7 +26,11 @@ import {
 } from "metabase/selectors/settings";
 import { getShowMetabaseLinks } from "metabase/selectors/whitelabel";
 import { Icon, type IconName } from "metabase/ui";
-import { type RecentItem, isRecentTableItem } from "metabase-types/api";
+import {
+  type RecentItem,
+  isRecentCollectionItem,
+  isRecentTableItem,
+} from "metabase-types/api";
 
 import type { PaletteAction } from "../types";
 import { filterRecentItems } from "../utils";
@@ -189,7 +193,7 @@ export const useCommandPalette = ({
                 trackSearchClick("item", index, "command-palette");
               },
               extra: {
-                isVerified: result.moderated_status === "verified",
+                moderatedStatus: result.moderated_status,
                 href: wrappedResult.getUrl(),
                 iconColor: icon.color,
                 subtext: getSearchResultSubtext(wrappedResult),
@@ -234,8 +238,9 @@ export const useCommandPalette = ({
           section: "recent",
           perform: () => {},
           extra: {
-            isVerified:
-              item.model !== "table" && item.moderated_status === "verified",
+            moderatedStatus: isRecentCollectionItem(item)
+              ? item.moderated_status
+              : null,
             href: Urls.modelToUrl(item),
             iconColor: icon.color,
             subtext: getRecentItemSubtext(item),
