+++ frontend/src/metabase/palette/types.ts
@@ -2,11 +2,12 @@ import type { LocationDescriptor } from "history";
 import type { Action, ActionImpl } from "kbar";
 
 import type { IconName } from "metabase/ui";
+import type { ModerationReviewStatus } from "metabase-types/api";
 
 interface PaletteActionExtras {
   extra?: {
     /** isVerified: If true, will show a verified badge next to the item name */
-    isVerified?: boolean;
+    moderatedStatus?: ModerationReviewStatus;
     /**
      * href: If defined, the palette item will be wrapped in a link. This allows for
      * browser interactions to open items in new tabs/windows
